+++ src/main/java/frc/robot/modules/swerveDrive/CTREConfigs.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.modules.swerveDrive;
 
 import com.ctre.phoenix.motorcontrol.SupplyCurrentLimitConfiguration;
 import com.ctre.phoenix.motorcontrol.can.TalonFXConfiguration;
@@ -6,6 +6,7 @@ import com.ctre.phoenix.sensors.AbsoluteSensorRange;
 import com.ctre.phoenix.sensors.CANCoderConfiguration;
 import com.ctre.phoenix.sensors.SensorInitializationStrategy;
 import com.ctre.phoenix.sensors.SensorTimeBase;
+import frc.robot.Constants;
 
 /**
  * CTRE config constants
