+++ src/main/java/frc/robot/modules/swerveDrive/CTREModuleState.java
@@ -1,4 +1,4 @@
-package frc.lib.util;
+package frc.robot.modules.swerveDrive;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
@@ -9,17 +9,14 @@ import edu.wpi.first.math.kinematics.SwerveModuleState;
 public class CTREModuleState {
 
     /**
-     * Minimize the change in heading the desired swerve module state would require by potentially
-     * reversing the direction the wheel spins. Customized from WPILib's version to include placing
-     * in appropriate scope for CTRE onboard control.
+     * Minimize the change in heading the desired swerve module state would require by potentially reversing the direction the wheel spins. Customized from WPILib's version to include placing in
+     * appropriate scope for CTRE onboard control.
      *
      * @param desiredState The desired state.
      * @param currentAngle The current module angle.
      */
-    public static SwerveModuleState optimize(SwerveModuleState desiredState,
-            Rotation2d currentAngle) {
-        double targetAngle = placeInAppropriate0To360Scope(currentAngle.getDegrees(),
-                desiredState.angle.getDegrees());
+    public static SwerveModuleState optimize(SwerveModuleState desiredState, Rotation2d currentAngle) {
+        double targetAngle = placeInAppropriate0To360Scope(currentAngle.getDegrees(), desiredState.angle.getDegrees());
         double targetSpeed = desiredState.speedMetersPerSecond;
         double delta = targetAngle - currentAngle.getDegrees();
         if (Math.abs(delta) > 90) {
