+++ src/main/java/frc/robot/subsystems/Swerve.java
@@ -11,7 +11,7 @@ import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.SwerveModule;
+import frc.robot.modules.swerveDrive.SwerveModule;
 
 /**
  * Creates swerve drive and commands for drive.
